Refactor api calls
